requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2+cu121; python_version >= "3.11"
torchvision==0.16.2+cu121; python_version >= "3.11"
torch==1.13.1+cu116; python_version == "3.10"
torchvision==0.14.1+cu116; python_version == "3.10"


hydra-core>=1.3.2